+++ expenso-api/src/auth/auth.service.hpp
@@ -3,7 +3,7 @@
 #include <variant>
 
 #include "auth/auth.models.hpp"
-#include "services/google-login.service.hpp"
+#include "services/googleLogin.service.hpp"
 #include "users/users.repository.hpp"
 #include "shared/models.hpp"
 #include "services/jwt.service.hpp"
