+++ expenso-api/src/services/mongoAccess.cpp
@@ -1,4 +1,4 @@
-#include "mongo-access.hpp"
+#include "mongoAccess.hpp"
 
 #include <mongocxx/logger.hpp>
 #include <mongocxx/stdx.hpp>
